requirements.txt
langchain
langchain-community
langchain-groq
unstructured
selenium
chromadb
streamlit
pandas
python-dotenv==0.19.2